Added a new task called SearchReplaceDb.
This new task uses the interconnectit database search / replace tool as
it's back end. For more info on this please see:

  - https://github.com/interconnectit/Search-Replace-DB

I am not 100% happy with the implementation. And I have plans to
completely rewrite the task using a more modern approach.
Using the Laravel Eloquent Query builder for the back end thus making
it work for any database, not just MySQL.

But for now this will do the job.