Updated to use our version of super_closure.
Okay so this is a little hacky I am the first to admit however I am lazy
and completely forgot to set the minimum-stability clause on another
project of mine which uses this one.

I tried for ages trying to get composer to update to the new version of
this package only the realise it was never going to update because we
were requiring an alpha package.